+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed script upgrade warnings for Unity 2022
 - Fixed ScreenManager editor texture to be applied on scene load
 - Added isQuest state to DebugState output of Sync Video Player
+- Added Global Video Tex option to CRT configurations
 
 ## [2.2.6] - 01-16-24